refactor: replace ReferenceLayersToggleControl with ReferenceLayersAndLocaleControl
